added json serializer for a record
added two methods on the Record class (pymarc/record.py)

1) ``as_dict(self)`` - which turns a record into a simplified dict,
   which will be used by ``as_json(self)``

2) ``as_json(self, **kwargs)`` - which serializes a record and returns a
   JSON string.

In order for this to work, python 2.6 is preferable because of the
stdlib's ``json`` module and some ``itertools`` functions. However,
I added some safety checks, whether ``json`` is actually available.
If it's not, the ``simplejson`` package will be used (available for python 2.3+).

On the ``itertools`` side, I added the necessary function
``izip_longest``.

``setup.py`` will check the python version and will add simplejson as
dependency if a version below 2.6 is used.

Added some basic tests in a new file: test/json_test.py

Tested on OS X with python versions 2.5, 2.6 and 2.7